# server/public/swagger.json
{
	"components": {
...
		"schemas": {
...
			"ModerationBlockUsersResponseDatum": {
				"properties": {
					"userId": {
						"type": "string"
					},
					"blockedAt": {
						"type": "string",
						"format": "date-time"
					},
					"user": {
						"$ref": "#/components/schemas/DatumUser"
					}
				},
				"required": [
					"userId",
					"blockedAt",
					"user"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ModerationBlockUsersResponse": {
				"properties": {
					"data": {
						"items": {
							"$ref": "#/components/schemas/ModerationBlockUsersResponseDatum"
						},
						"type": "array"
					}
				},
				"required": [
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ModerationBlockUsersRequest": {
				"properties": {
					"blockedUserId": {
						"type": "string"
					}
				},
				"required": [
					"blockedUserId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ProfilePicture": {
				"properties": {
					"url": {
						"type": "string"
					},
					"width": {
						"type": "number",
						"format": "double"
					},
					"height": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"url",
					"width",
					"height"
				],
				"type": "object",
				"additionalProperties": false
			},
			"FeedsFriendsResponseUser": {
				"properties": {
					"id": {
						"type": "string"
					},
					"username": {
						"type": "string"
					},
					"profilePicture": {
						"$ref": "#/components/schemas/ProfilePicture"
					}
				},
				"required": [
					"id",
					"username"
				],
				"type": "object",
				"additionalProperties": false
			},
			"MediaType": {
				"enum": [
					"late",
					"photo"
				],
				"type": "string"
			},
			"Region": {
				"enum": [
					"europe-west",
					"us-central"
				],
				"type": "string"
			},
			"Bucket": {
				"enum": [
					"storage.bere.al"
				],
				"type": "string"
			},
			"PostLocation": {
				"properties": {
					"_latitude": {
						"type": "number",
						"format": "double"
					},
					"_longitude": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"_latitude",
					"_longitude"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreationDate": {
				"properties": {
					"_seconds": {
						"type": "number",
						"format": "double"
					},
					"_nanoseconds": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"_seconds",
					"_nanoseconds"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"FeedsFriendsResponseRealMoji": {
				"properties": {
					"id": {
						"type": "string"
					},
					"uid": {
						"type": "string"
					},
					"userName": {
						"type": "string"
					},
					"user": {
						"$ref": "#/components/schemas/PostUser"
					},
					"emoji": {
						"type": "string"
					},
					"type": {
						"$ref": "#/components/schemas/Type"
					},
					"uri": {
						"type": "string"
					},
					"date": {
						"$ref": "#/components/schemas/CreationDate"
					}
				},
				"required": [
					"id",
					"uid",
					"userName",
					"user",
					"emoji",
					"type",
					"uri",
					"date"
				],
				"type": "object",
				"additionalProperties": false
			},
			"FeedsFriendsResponse": {
				"properties": {
					"id": {
						"type": "string"
					},
					"notificationID": {
						"type": "string"
					},
					"ownerID": {
						"type": "string"
					},
					"userName": {
						"type": "string"
					},
					"user": {
						"$ref": "#/components/schemas/FeedsFriendsResponseUser"
					},
					"mediaType": {
						"$ref": "#/components/schemas/MediaType"
					},
					"region": {
						"$ref": "#/components/schemas/Region"
					},
					"bucket": {
						"$ref": "#/components/schemas/Bucket"
					},
					"photoURL": {
						"type": "string"
					},
					"imageWidth": {
						"type": "number",
						"format": "double"
					},
					"imageHeight": {
						"type": "number",
						"format": "double"
					},
					"secondaryPhotoURL": {
						"type": "string"
					},
					"secondaryImageHeight": {
						"type": "number",
						"format": "double"
					},
					"secondaryImageWidth": {
						"type": "number",
						"format": "double"
					},
					"members": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"lateInSeconds": {
						"type": "number",
						"format": "double"
					},
					"isPublic": {
						"type": "boolean"
					},
					"location": {
						"$ref": "#/components/schemas/PostLocation"
					},
					"caption": {
						"type": "string"
					},
					"retakeCounter": {
						"type": "number",
						"format": "double"
					},
					"creationDate": {
						"$ref": "#/components/schemas/CreationDate"
					},
					"updatedAt": {
						"type": "number",
						"format": "double"
					},
					"takenAt": {
						"$ref": "#/components/schemas/CreationDate"
					},
					"comment": {
						"items": {},
						"type": "array"
					},
					"realMojis": {
						"items": {
							"$ref": "#/components/schemas/FeedsFriendsResponseRealMoji"
						},
						"type": "array"
					},
					"screenshots": {
						"items": {},
						"type": "array"
					},
					"screenshotsV2": {
						"items": {},
						"type": "array"
					},
					"visibility": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"notificationID",
					"ownerID",
					"userName",
					"user",
					"mediaType",
					"region",
					"bucket",
					"photoURL",
					"imageWidth",
					"imageHeight",
					"secondaryPhotoURL",
					"secondaryImageHeight",
					"secondaryImageWidth",
					"members",
					"lateInSeconds",
					"isPublic",
					"retakeCounter",
					"creationDate",
					"updatedAt",
					"takenAt",
					"comment",
					"realMojis",
					"screenshots",
					"screenshotsV2"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"FeedsDiscoveryResponse": {
				"properties": {
					"posts": {
						"items": {
							"$ref": "#/components/schemas/Post"
						},
						"type": "array"
					},
					"lastIndex": {
						"type": "string"
					}
				},
				"required": [
					"posts",
					"lastIndex"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"FeedsMemoriesResponseDatum": {
				"properties": {
					"id": {
						"type": "string"
					},
					"thumbnail": {
						"$ref": "#/components/schemas/ProfilePicture"
					},
					"primary": {
						"$ref": "#/components/schemas/ProfilePicture"
					},
					"secondary": {
						"$ref": "#/components/schemas/ProfilePicture"
					},
					"isLate": {
						"type": "boolean"
					},
					"memoryDay": {
						"type": "string",
						"format": "date-time"
					},
					"location": {
						"allOf": [
							{
								"$ref": "#/components/schemas/DatumLocation"
							}
						],
						"nullable": true
					}
				},
				"required": [
					"id",
					"thumbnail",
					"primary",
					"secondary",
					"isLate",
					"memoryDay",
					"location"
				],
				"type": "object",
				"additionalProperties": false
			},
			"FeedsMemoriesResponse": {
				"properties": {
					"data": {
						"items": {
							"$ref": "#/components/schemas/FeedsMemoriesResponseDatum"
						},
						"type": "array"
					},
					"next": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					},
					"memoriesSynchronized": {
						"type": "boolean"
					}
				},
				"required": [
					"data",
					"next",
					"memoriesSynchronized"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SendVerificationCodeResponse": {
				"properties": {
					"sessionInfo": {
						"type": "string"
					}
				},
				"required": [
					"sessionInfo"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Pick_SendVerificationCodeRequest.Exclude_keyofSendVerificationCodeRequest.iosReceipt__": {
				"properties": {
					"phoneNumber": {
						"type": "string"
					}
				},
				"required": [
					"phoneNumber"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"Omit_SendVerificationCodeRequest.iosReceipt_": {
				"$ref": "#/components/schemas/Pick_SendVerificationCodeRequest.Exclude_keyofSendVerificationCodeRequest.iosReceipt__",
				"description": "Construct a type with the properties of T except for those in type K."
			},
			"VerifyPhoneNumberResponse": {
				"properties": {
					"idToken": {
						"type": "string"
					},
					"refreshToken": {
						"type": "string"
					},
					"expiresIn": {
						"type": "string"
					},
					"localId": {
						"type": "string"
					},
					"isNewUser": {
						"type": "boolean"
					},
					"phoneNumber": {
						"type": "string"
					}
				},
				"required": [
					"idToken",
					"refreshToken",
					"expiresIn",
					"localId",
					"isNewUser",
					"phoneNumber"
				],
				"type": "object",
				"additionalProperties": false
			},
			"VerifyPhoneNumberRequest": {
				"properties": {
					"sessionInfo": {
						"type": "string"
					},
					"code": {
						"type": "string"
					},
					"operation": {
						"type": "string"
					}
				},
				"required": [
					"sessionInfo",
					"code",
					"operation"
				],
				"type": "object",
				"additionalProperties": false
			},
			"RefreshTokenResponse": {
				"properties": {
					"access_token": {
						"type": "string"
					},
					"expires_in": {
						"type": "string"
					},
					"token_type": {
						"type": "string"
					},
					"refresh_token": {
						"type": "string"
					},
					"id_token": {
						"type": "string"
					},
					"user_id": {
						"type": "string"
					},
					"project_id": {
						"type": "string"
					}
				},
				"required": [
					"access_token",
					"expires_in",
					"token_type",
					"refresh_token",
					"id_token",
					"user_id",
					"project_id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"RefreshTokenRequest": {
				"properties": {
					"refresh_token": {
						"type": "string"
					},
					"grant_type": {
						"type": "string"
					}
				},
				"required": [
					"refresh_token",
					"grant_type"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Realmoji": {
				"properties": {
					"emoji": {
						"type": "string"
					},
					"media": {
						"$ref": "#/components/schemas/ProfilePicture"
					}
				},
				"required": [
					"emoji",
					"media"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"Stats": {
				"properties": {
					"postTotal": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"postTotal"
				],
				"type": "object",
				"additionalProperties": false
			},
			"MeResponse": {
				"properties": {
					"id": {
						"type": "string"
					},
					"username": {
						"type": "string"
					},
					"birthdate": {
						"type": "string",
						"format": "date-time"
					},
					"fullname": {
						"type": "string"
					},
					"profilePicture": {
						"$ref": "#/components/schemas/ProfilePicture"
					},
					"realmojis": {
						"items": {
							"$ref": "#/components/schemas/Realmoji"
						},
						"type": "array"
					},
					"devices": {
						"items": {
							"$ref": "#/components/schemas/Device"
						},
						"type": "array"
					},
					"stats": {
						"$ref": "#/components/schemas/Stats"
					},
					"canDeletePost": {
						"type": "boolean"
					},
					"canUpdateRegion": {
						"type": "boolean"
					},
					"phoneNumber": {
						"type": "string"
					},
					"biography": {
						"type": "string"
					},
					"location": {
						"type": "string"
					},
					"countryCode": {
						"type": "string"
					},
					"region": {
						"$ref": "#/components/schemas/Region"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"newUsername": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"username",
					"birthdate",
					"fullname",
					"profilePicture",
					"realmojis",
					"devices",
					"stats",
					"canDeletePost",
					"canUpdateRegion",
					"phoneNumber",
					"biography",
					"location",
					"countryCode",
					"region",
					"createdAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PersonMePatchRequest": {
				"properties": {
					"biography": {
						"type": "string"
					}
				},
				"required": [
					"biography"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PersonMePostRequest": {
				"properties": {
					"username": {
						"type": "string"
					},
					"birthdate": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"username",
					"birthdate"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PersonMeUsernamePatchRequest": {
				"properties": {
					"username": {
						"type": "string"
					}
				},
				"required": [
					"username"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"RelationshipsFriendsResponse": {
				"properties": {
					"data": {
						"items": {
							"$ref": "#/components/schemas/RelationshipsFriendsResponseDatum"
						},
						"type": "array"
					},
					"next": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					}
				},
				"required": [
					"data",
					"next"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"RelationshipsFriendRequestsSentResponse": {
				"properties": {
					"data": {
						"items": {
							"$ref": "#/components/schemas/RelationshipsFriendRequestsSentResponseDatum"
						},
						"type": "array"
					},
					"next": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					}
				},
				"required": [
					"data",
					"next"
				],
				"type": "object",
				"additionalProperties": false
			},
			"RelationshipsSuggestionsResponse": {
				"properties": {
					"data": {
						"items": {},
						"type": "array"
					},
					"next": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					}
				},
				"required": [
					"data",
					"next"
				],
				"type": "object",
				"additionalProperties": false
			},
			"RelationshipsFriendRequestsPostResponse": {
				"properties": {
					"statusCode": {
						"type": "number",
						"format": "double"
					},
					"errorKey": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					},
					"fields": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					},
					"timestamp": {
						"type": "string",
						"format": "date-time"
					},
					"requestId": {
						"type": "string"
					}
				},
				"required": [
					"statusCode",
					"errorKey",
					"fields",
					"timestamp",
					"requestId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"RelationshipsFriendRequestsPostRequest": {
				"properties": {
					"userId": {
						"type": "string"
					},
					"source": {
						"type": "string"
					}
				},
				"required": [
					"userId",
					"source"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "server",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/api/moderation/block-users": {
			"get": {
				"operationId": "GetBlockedUsers",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ModerationBlockUsersResponse"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "header",
						"name": "authorization",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"post": {
				"operationId": "PostBlockUser",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "header",
						"name": "authorization",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ModerationBlockUsersRequest"
							}
						}
					}
				}
			}
		},
		"/api/moderation/block-users/{userId}": {
			"delete": {
				"operationId": "DeleteBlockUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "header",
						"name": "authorization",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/feeds/friends": {
			"get": {
				"operationId": "GetFriends",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/FeedsFriendsResponse"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "header",
						"name": "authorization",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/feeds/discovery": {
			"get": {
				"operationId": "GetDiscovery",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/FeedsDiscoveryResponse"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "header",
						"name": "authorization",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/feeds/memories": {
			"get": {
				"operationId": "GetMemories",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/FeedsMemoriesResponse"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "header",
						"name": "authorization",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/feeds/memories/{memoryId}": {
			"delete": {
				"operationId": "DeleteMemory",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/FeedsMemoriesResponse"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "header",
						"name": "authorization",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "memoryId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/login/sendVerificationCode": {
			"post": {
				"operationId": "PostSendVerificationCode",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SendVerificationCodeResponse"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Omit_SendVerificationCodeRequest.iosReceipt_"
							}
						}
					}
				}
			}
		},
		"/api/login/verifyPhoneNumber": {
			"post": {
				"operationId": "PostVerifyPhoneNumber",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/VerifyPhoneNumberResponse"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/VerifyPhoneNumberRequest"
							}
						}
					}
				}
			}
		},
		"/api/login/refreshToken": {
			"post": {
				"operationId": "PostRefreshToken",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RefreshTokenResponse"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/RefreshTokenRequest"
							}
						}
					}
				}
			}
		},
		"/api/person/me": {
			"get": {
				"operationId": "GetMe",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MeResponse"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "header",
						"name": "authorization",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"patch": {
				"operationId": "PatchMe",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MeResponse"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "header",
						"name": "authorization",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/PersonMePatchRequest"
							}
						}
					}
				}
			},
			"post": {
				"operationId": "PostMe",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "header",
						"name": "authorization",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/PersonMePostRequest"
							}
						}
					}
				}
			}
		},
		"/api/person/me/username": {
			"patch": {
				"operationId": "PatchMeUsername",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MeResponse"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "header",
						"name": "authorization",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/PersonMeUsernamePatchRequest"
							}
						}
					}
				}
			}
		},
		"/api/relationships/friends": {
			"get": {
				"operationId": "GetFriends",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RelationshipsFriendsResponse"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "header",
						"name": "authorization",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/relationships/friend-requests/sent": {
			"get": {
				"operationId": "GetFriendRequestsSent",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RelationshipsFriendRequestsSentResponse"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "header",
						"name": "authorization",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/relationships/friend-requests/received": {
			"get": {
				"operationId": "GetFriendRequestsReceived",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RelationshipsSuggestionsResponse"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "header",
						"name": "authorization",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/relationships/suggestions": {
			"get": {
				"operationId": "GetSuggestions",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RelationshipsSuggestionsResponse"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "header",
						"name": "authorization",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/relationships/friend-requests": {
			"post": {
				"operationId": "PostFriendRequest",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RelationshipsFriendRequestsPostResponse"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "header",
						"name": "authorization",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/RelationshipsFriendRequestsPostRequest"
							}
						}
					}
				}
			}
		},
		"/api/relationships/contacts": {
			"post": {
				"operationId": "PostContacts",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "header",
						"name": "authorization",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"items": {
									"type": "string"
								},
								"type": "array"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}